requirements.txt
flask==3.0.0
mysql-connector-python==8.1.0
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.24.3
python-dotenv==1.0.0
scikit-learn==1.2.2
sentence-transformers==2.2.2
transformers==4.30.2
torch==2.0.1
huggingface-hub==0.15.1
pytest==7.4.0
requests-mock==1.11.0
faiss-cpu==1.8.0